backend/requirements.txt
# Flask and CORS for API
flask==2.3.2
flask-cors==3.0.10

# TensorFlow for training and prediction
tensorflow==2.10.0

# YOLOv5 Dependencies
torch==1.12.0                # PyTorch for YOLOv5
ultralytics==8.0.0           # YOLOv5/YOLOv8 model
opencv-python-headless==4.6.0.66  # OpenCV for image processing

# Array and image manipulation
numpy==1.23.5
pillow==9.3.0

# Visualization
matplotlib==3.6.2
